Round percentage back to int to avoid displaying decimals when using %p
`std::to_string`, when used on doubles, will print out the double with 6 digits of precision (eg: 95 is printed as 95.000000), which looks ugly. This can technically be altered using `setprecision` (http://en.cppreference.com/w/cpp/io/manip/setprecision), but that seems really annoying. Instead, we can just cast back to an integer, which should print just fine.

+++ src/Judgments.cpp
@@ -110,7 +110,7 @@ GetJudgementText(Judgement& judgement, int score, int before, int after, int acc
     text.set_accuracy(std::to_string(accuracy));
     text.set_afterCut(std::to_string(after));
     text.set_score(std::to_string(score));
-    text.set_percent(std::to_string(round(100 * (float) score / maxScore)));
+    text.set_percent(std::to_string((int)round(100 * (float) score / maxScore)));
     text.set_timeDependency(TimeDependenceString(timeDependence));
     text.set_beforeCutSegment(GetBestSegmentText(getGlobalConfig().CurrentConfig.BeforeCutAngleSegments, before));
     text.set_accuracySegment(GetBestSegmentText(getGlobalConfig().CurrentConfig.AccuracySegments, accuracy));
